+++ packages/component/src/useSpeak.js
@@ -1,8 +1,0 @@
-import { useContext } from 'react';
-import Context from './Context';
-
-export default function useSpeak() {
-  const { ponyfill, speak } = useContext(Context);
-
-  return (utterance, progressFn) => speak(ponyfill, utterance, { onStart: progressFn && (() => progressFn()) });
-}