+++ CHANGELOG.md
@@ -13,15 +13,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
    - Support nested `<Context>` and `<Composer>`
       - Nested context will share the same queue, with different ponyfills
       - Upgraded `<Composer>` component from class component to functional component
-   - Added new `useSpeak()` hook, which returns a `speak(utterance, progressFn)` function to queue an utterance
+   - Added new `useSynthesize()` hook, which returns a function to queue an utterance
+      - `synthesize(utteranceOrText: (SpeechSynthesisUtterance | string), progressFn: () => void) => void`
    - New `<SayUtterance>` component to synthesis `SpeechSynthesisUtterance` instead of text
       - `<Say>` is being refactored to use `<SayUtterance>` to simplify the codebase
       - `<SayButton>` is being refactored to use `<Say>` to simplify the codebase
    - `Composer` signature is being updated
-      - `speak` no longer requires an `id` and will return `{ cancel: function, promise: Promise }`
+      - New `synthesize` function to replace `speak` and `cancel`
+         - When called, it will return `{ cancel: Function, promise: Promise }`
          - `cancel`, when called, will cancel the utterance. If the utterance is being synthesized, it will be stopped abruptly
          - `promise` will be resolved when the utterance is synthesized or errored
-      - `cancel` is removed because `speak` now expose a `cancel` function, which offer same functionality
+      - `cancel` and `speak` is removed because the newer `synthesize` function offer same functionality with simplified interface
    - All React components now accepts `ponyfill` instead of `speechSynthesis` and `speechSynthesisUtterance`
       - Using browser speech would become as simple as `<Say ponyfill={ window }>`